requirements.txt
asgiref==3.4.1
autopep8==2.0.0
Django==3.2.16
django-autoslug==1.9.8
django-cities-light==3.9.2
django-cors-headers==3.10.1
django-crispy-forms==1.13.0
django-online-users==0.3
django-registration-redux==2.11
django-smart-selects==1.6.0
djangorestframework==3.14.0
Pillow==8.4.0
progressbar2==3.55.0
pycodestyle==2.9.1
python-utils==3.3.3
pytz==2022.5
six==1.16.0
sqlparse==0.4.3
tomli==1.2.3
typing_extensions==4.1.1
Unidecode==1.3.6
